+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [actions-package] github issue# 73 is addressed. Solution is to replace line breaks in description by space
 * [artifact-manager] Exporting vROps dashboards fails if metadata folder is not created
 * [artifact-manager] 108 / Pushing and pulling from/to vRA 8.12 fails with buildtools 2.32.0
+* [vRA-NG] IAC-772 / When pushing Resource Actions (RAs), the ones that are scoped to a single project, should have their projectId modified to correspond to the one in configuration
 
 ## v2.33.0 - 02 Jun 2023